pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "s1"
version = "0.1.0"
description = "Only a litle bit of S3."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.24,<0.30",
    "google-cloud-storage>=2.13,<3.0",
    "pyarrow>=15.0,<16.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0,<9.0",
    "minio>=7.0,<8.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
py-modules = ["main"]
